+++ spec/spec_helper.rb
@@ -4,6 +4,10 @@ require 'rspec'
 require 'stripe'
 require 'rack/test'
 
+project_root = File.expand_path("..", File.dirname(__FILE__))
+$LOAD_PATH << project_root
+Dir.glob("spec/support/**/*.rb").each { |file| require file }
+
 RSpec.configure do |config|
   config.expect_with :rspec do |c|
     c.syntax = :expect
@@ -15,3 +19,6 @@ RSpec.configure do |config|
 
   config.order = 'random'
 end
+
+# Do not announce that capybara is starting puma
+Capybara.server = :puma, { Silent: true }